Convert all absence routes to controllers
